Combine the netframework_clr* collectors into a single collector
The new collector is named netframework_clr and supports a configurable
set of child collectors, similar to the mssql collector. Other changes
include:
* Perflib is used instead of WMI.
* Process name whitelist and blacklist flags are added to
  specify regexp to use for matching for inclusion and exclusion,
  respectively.

There is no change to metric names or labels.
