+++ docs/README.md
@@ -15,14 +15,7 @@ This directory contains documentation of the collectors in the windows_exporter,
 - [`memory`](collector.memory.md)
 - [`msmq`](collector.msmq.md)
 - [`mssql`](collector.mssql.md)
-- [`netframework_clrexceptions`](collector.netframework_clrexceptions.md)
-- [`netframework_clrinterop`](collector.netframework_clrinterop.md)
-- [`netframework_clrjit`](collector.netframework_clrjit.md)
-- [`netframework_clrloading`](collector.netframework_clrloading.md)
-- [`netframework_clrlocksandthreads`](collector.netframework_clrlocksandthreads.md)
-- [`netframework_clrmemory`](collector.netframework_clrmemory.md)
-- [`netframework_clrremoting`](collector.netframework_clrremoting.md)
-- [`netframework_clrsecurity`](collector.netframework_clrsecurity.md)
+- [`netframework_clr`](collector.netframework_clr.md)
 - [`net`](collector.net.md)
 - [`os`](collector.os.md)
 - [`process`](collector.process.md)
